Enabling Gemini as backend for Rhesis metrics
