+++ apps/backend/pyproject.toml
@@ -44,7 +44,9 @@ dependencies = [
     "gunicorn>=21.2.0",
     "tomli>=2.0.0",
     "google-genai==1.13.0",
-    "mirascope==1.24.0"
+    "mirascope==1.24.0",
+    "protobuf>=5.29.4",
+    "proto-plus>=1.26.1",
 ]
 
 [build-system]
